relocateNegatives moved to functions.h
