+++ CMakeLists.txt
@@ -5,6 +5,7 @@ project(radix-sort)
 include_directories(include)
 
 set(RADIX_SORT_SOURCE
+    src/functions.cpp
     src/main.cpp
     src/recursivesequentialsort.cpp
     src/sequentialsort.cpp
