+++ include/functions.h
@@ -1,0 +1,9 @@
+#pragma once
+
+/**
+ * @brief Function to swap negative elements of arrat with positive
+ * 
+ * @param first array last element pointer
+ * @param last  order of compared digit
+ */
+void relocateNegatives(int* first, int* last);
