+++ include/recursivesequentialsort.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <algorithm>
 #include <iterator>
 
 /**
@@ -19,11 +18,3 @@ void recursiveSequentialSort(int* first, int* last);
  * @param order order of compared digit
  */
 void recursiveSequentialSortStage(int* first, int* last, size_t order);
-
-/**
- * @brief Function to swap negative elements of arrat with positive
- * 
- * @param first array last element pointer
- * @param last  order of compared digit
- */
-void relocateNegatives(int* first, int* last);
