+++ src/recursivesequentialsort.cpp
@@ -1,14 +1,8 @@
 #include "recursivesequentialsort.h"
 
-
 void recursiveSequentialSort(int* first, int* last)
 {
-    int min = *std::min_element(first, last);
     recursiveSequentialSortStage(first, last, sizeof(int) * 8);
-    if (min < 0)
-    {
-        relocateNegatives(first, last);
-    }
 }
 
 void recursiveSequentialSortStage(int* first, int* last, size_t order)
@@ -44,28 +38,4 @@ void recursiveSequentialSortStage(int* first, int* last, size_t order)
     }
     recursiveSequentialSortStage(first, end, order - 1);
     recursiveSequentialSortStage(begin, last, order - 1);
-}
-
-void relocateNegatives(int* first, int* last)
-{
-    size_t size = last - first + 1;
-    int* result = new int[size];
-    int* separator = first;
-    for (; (separator <= last) && !(*separator < 0); separator++)
-    {
-
-    }
-    size_t i = 0;
-    for (int* j = separator; j <= last; j++, i++)
-    {
-        result[i] = *j;
-    }
-    for (int* j = first; j < separator; j++, i++)
-    {
-        result[i] = *j;
-    }
-    for (size_t k = 0; k < size; k++)
-    {
-        first[k] = result[k];
-    }
 }
